==> scripts/food_tree/scripts/03_compare_computrition_and_fndds.Rmd <==
---
title: "Compare computrition with FNDDS value"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(ggpubr)
library(kableExtra)
``` 

```{r}
ALL <- read_csv('../data/finalized/all_patients_record.csv')

# the nutrient value for all the Food_NSC
uniq_fndds <- read_csv('../data/finalized/uniq_Food_NSC_fndds_nutrient_value.csv') %>% 
   gather('item', 'value_fndds', names(.)[2]:names(.)[ncol(.)])
```

```{r}
# now collapse the df and merge the two into a single df
computrition <- ALL %>% 
  select(Food_NSC,Calories_kcal: Sodium_mg) %>% 
  distinct(Food_NSC, .keep_all = T) %>% 
  dplyr::rename(energy = Calories_kcal,
         protein = Proteing_g,
         fat = Fat_g,
         carbohydrate = Carbohydrates_g,
         fiber = Fibers_g,
         sugar = Sugars_g, 
         sodium = Sodium_mg) %>% 
  gather('item', 'value', names(.)[2]:names(.)[ncol(.)])

com_colnam <- colnames(computrition)


combined <- computrition %>% 
  inner_join(uniq_fndds)
```

```{r}
# look at the items with energy == 0
energy0 <- combined %>% 
  filter(item == 'energy' & value == 0) # some of them are suspicious such as the condiments one


energy0 %>% 
  write_csv('../data/source/energy0_item.csv')
```


```{r}
# look at the computrition energy != 0 first 
check <- combined %>% 
  split(.$item == 'energy' & .$value == 0) %>% 
  pluck('FALSE')

ratio_cal <- check %>% 
  filter(item == 'energy' & value != 0) %>% 
  mutate(ratio = round(value_fndds/value, 2)) %>% 
  select(Food_NSC, ratio)

compare2 <- check %>% 
  left_join(ratio_cal , by = 'Food_NSC') %>% 
  mutate(computri_converted = value*ratio) 
```

```{r}
# now compare the six nutrients
compare2_res <- compare2 %>% 
  filter(item != 'energy') %>% 
  filter(!is.na(computri_converted)) %>% 
  rename(value_compu = value) %>% 
  split(.$item)
```


```{r}
compare2_res %>% 
  pluck('fiber') %>% 
  head() %>% 
  kable() %>% 
  kable_styling(full_width = F)

compare2_res %>% 
  pluck('carbohydrate') %>% 
  head() %>% 
  kable() %>% 
  kable_styling(full_width = F)

check %>% 
  filter(item == 'energy' & value != 0) %>% 
  mutate(ratio = round(value_fndds/value, 2)) %>% 
  rename(value_compu = value) %>% 
  head() %>% 
  kable() %>% 
  kable_styling(full_width = F)
```

```{r}
compare2_pearson <- compare2_res %>% 
  map_dfr(~ round(cor.test(.$value_fndds, 
                           .$computri_converted, 
                           method = 'pearson', 
                           exact = F)$estimate, 2)) %>% 
  gather(key = 'nutrient', value = 'pearson_cor') %>% 
  arrange(pearson_cor)  



# get the pvalue and the correlation and add to the scatter plot

compare2_pearson_pval <- compare2_res %>% 
  map_dfr(~ round(cor.test(.$value_fndds, 
                           .$computri_converted, 
                           method = 'pearson', 
                           exact = F)$p.value, 2)) %>% 
  gather(key = 'nutrient', value = 'pearson_pval') %>% 
  arrange(pearson_pval) 

person_annot <- compare2_pearson_pval %>% 
  full_join(compare2_pearson) %>% 
  split(.$nutrient)
names(person_annot)
names(compare2_res)
```

```{r}
# make scatter plot to visualize the correlation
res_plot <- list()

for (i in 1:length(compare2_res)) {
 res_plot[[i]] <-  compare2_res %>% 
    pluck(names(compare2_res)[i]) %>% 
    ggscatter('value_fndds', 'computri_converted',
              title = str_glue('FNDDS value and converted computrition value\n{names(compare2_res)[i]}\nPearson cor: {person_annot[[i]]$pearson_cor}, pvalue: {person_annot[[i]]$pearson_pval}'),
              xlab = 'FNDDS value per 100g edible portion',
              ylab = 'Computrition value per 100g edible portion',
              alpha = 0.5,
              add.params = list(size = 4)) +
    geom_abline(intercept = 0, slope = 1, col = 'red', linetype = 'dashed') 
}

#+ggsave(str_glue('../figs/correlation_scatter_{names(compare2_res)[i]}.jpg'), width = 6, height = 5)

library(cowplot)
g <- plot_grid(res_plot[[1]],res_plot[[2]],res_plot[[3]],res_plot[[4]],res_plot[[5]],res_plot[[6]],
  nrow = 2,
  align = 'hv',
  #labels = 'AUTO',
  axis = 'b') +
  ggsave('../figs/scatter_all.jpg', width = 12, height = 8, dpi = 300)

```

```{r}
compare2_spearman <- compare2_res %>% 
  map_dfr(~ round(cor.test(.$value_fndds, 
                           .$computri_converted, 
                           method = 'spearman', 
                           exact = F)$estimate, 2)) %>% 
  gather(key = 'nutrient', value = 'spearman_cor') %>% 
  arrange(spearman_cor)  


cor_all <- compare2_spearman %>% 
  full_join(compare2_pearson, by = 'nutrient') %>% 
  arrange(pearson_cor)

cor_all %>% 
  write_csv('../data/corerlation_nutrients_foods.csv')
  
cor_all %>% 
  kable() %>% 
  kable_styling(full_width = F)
```

```{r}
doubt <- compare2 %>% 
  filter(item != 'energy') %>% 
  filter(is.na(computri_converted)) 
```


```{r}
# find the outlier of the scatter plots comparison by doing a ratio of the converted computrition value and FNDDS value
outlier <- compare2_res %>% 
  map(~ mutate(.data = ., check_ratio = round(computri_converted/value_fndds, 2))) %>% 
  bind_rows(.id = 'group') %>% 
  filter(is.finite(check_ratio))
```


```{r}
# check the ratio much bigger than 1 ones 
# >= 4
outlier_select <- outlier %>% 
  filter(check_ratio >= 4) %>% 
  arrange(Food_NSC) %>% 
  distinct(Food_NSC)


ALL %>% 
  distinct(Food_NSC, .keep_all = T) %>% 
  filter(Food_NSC %in% outlier_select$Food_NSC) %>% 
  write_csv('../data/outlier_very_obvious.csv')
```


```{r}
# the food with ration >= 4 with all the macronutrients except sodium for peter to check
ratio4 <- ALL %>% 
  filter(Food_NSC %in% outlier_select$Food_NSC) %>% 
  select(Food_NSC: Sugars_g) %>% 
  distinct()

ratio4 %>% 
  write_csv('../data/need_to_be_checked_comput_fndds_ratio_too_big_all_columns.csv')
```


```{r}
# retrieve the ID of the food and join back to this table
outlier_select <- outlier_select %>% 
  left_join(ALL %>%  select(key, Food_NSC)) %>% 
  select(key, group : check_ratio)

outlier_select %>% 
  write_csv('../data/need_to_be_checked_comput_fndds_ratio_too_big.csv')
```


```{r}
# or I just pick the points that are in the impossible region
impossible_all <- compare2_res %>% 
  bind_rows(.id = 'group') 


carbohydrate <- impossible_all%>% 
  filter(group == 'carbohydrate' & computri_converted > 100)

fat <- impossible_all%>% 
  filter(group == 'fat' & computri_converted > 101)

fiber <- impossible_all %>% 
  filter(group == 'fiber' & computri_converted > 20)


sodium <- impossible_all%>% 
  filter(group == 'sodium'  & computri_converted > 5000 & value_fndds < 3000)

all_obvious <- bind_rows(carbohydrate, fat, fiber, sodium)


all_obvious %>% 
  left_join(ALL %>%  select(key, Food_NSC), by = 'Food_NSC') %>% 
  write_csv('../data/outlier_very_obvious.csv')

correct_id <- all_obvious %>% 
  left_join(ALL %>%  select(key, Food_NSC))

ALL %>% 
  distinct(Food_NSC, .keep_all = T) %>% 
  filter(Food_NSC %in% all_obvious$Food_NSC) %>% 
  write_csv('../data/outlier_very_obvious.csv')
```
