==> scripts/food_tree/scripts/04_explore_nutrient_foodgroup.Rmd <==
---
title: "Explore the ALL records, on a nutrient and food group level"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(ggpubr)
library(kableExtra)
library(RColorBrewer)
library(readxl)
library(lubridate)
source('~/db.R')
```

```{r}
# loading the finalized table
Final <- read_csv('../data/finalized/all_patients_record_w_food_code.csv')
```

```{r}
# the date of engraftment
# transplant type, intensity, etc. The column titled ‘tt_anc_500’ is the day post -HCT of engraftment and ‘anc_500’ is the specific date of engraftment.
current_mrns <- Final %>% 
           distinct(mrn) %>% pull(mrn)

engra <- read_csv('../data/source/nutrition_engraftment day_02.26.20.csv') %>% 
  filter(mrn %in% current_mrns) %>% 
  arrange(hct) %>% 
  distinct(mrn, .keep_all = T) %>% 
  mutate(dayEngra = anc_500 - hct)
```


```{r}
# the EN TPN data
tpn <- read_csv('../data/source/TPN_EN patients.csv') %>% 
  mutate(Date = mdy(Date))

# find the hct of those patients
TPN <- tpn %>% 
  select(mrn, Date, type= TPN.1..EN.2) %>% 
  inner_join(engra %>% 
               select(mrn, hct))  %>% 
  mutate(dayRT = Date-hct) %>% 
  mutate(type = if_else(type ==1, "TPN", "EN"))

TPN %>% 
  write_csv('../data/finalized/TPN_EN_all.csv')

```
## The food group for each patient 

```{r}
# node labels different levels of nodes
# setting the color for each food group
nodes <- read_tsv('../data/source/NodeLabels.txt', col_types = 'cc') %>% 
  filter(nchar(Level.code) == 1) %>% 
  rename(food_grp = Level.code)

scale_fill_manual_value <- c("Milk_and_Milk_Products" = "#a9a9a9", "Meat_Poultry_Fish_and_Mixtures" = "#800000",
                               "Eggs" = "#ffe119", "Dry_Beans_Peas_Other_Legumes_Nuts_and_Seeds" = "#808000",
                               "Grain_Product" = "#9a6324", "Fruits" = "#911eb4",
                               "Vegetables" = "#469990", "Fats_Oils_and_Salad_Dressings" = "#000075", "Sugars_Sweets_and_Beverages" = "#e5a47c")

nodes %>% 
  kable() %>% 
  kable_styling(full_width = F)
```

```{r}
# getting the food group of each food
fgroup <- Final %>% 
  select(mrn, DayRT, dehydrated_weight, Food_code) %>% 
  mutate(food_grp = str_sub(Food_code, 1, 1))

# the total dehydated weight for a patient for one day
total_daily <- fgroup %>% 
  group_by(mrn, DayRT) %>% 
  summarise(daily_tol = sum(dehydrated_weight))

# the total dehydated weight for a patient for one day for a single food group
food_per_group <- fgroup %>% 
  group_by(mrn, DayRT, food_grp) %>% 
  summarise(grp_tol = sum(dehydrated_weight))

# joined together
fgroup_relab <- food_per_group %>% 
  left_join(total_daily) %>% 
  mutate(grp_relab = grp_tol/daily_tol) %>% 
  arrange(mrn, DayRT, food_grp) %>% 
  ungroup() %>% 
  left_join(nodes, by = 'food_grp') %>% 
  mutate(Main.food.description = factor(Main.food.description, levels = c('Grain_Product','Sugars_Sweets_and_Beverages','Meat_Poultry_Fish_and_Mixtures','Milk_and_Milk_Products','Vegetables','Fruits','Dry_Beans_Peas_Other_Legumes_Nuts_and_Seeds','Eggs','Fats_Oils_and_Salad_Dressings'))) %>% 
  # if one whole day the pt only has consumed suger beverage the value is NaN and need to set that to be 1 for the group 9
  mutate(grp_relab = if_else(is.nan(grp_relab), 1, grp_relab))

# daily_tol is the total dehydated weight for a patient for one day
# grp_tol is the total dehydated weight for a patient for one day for a single food group

fgroup_relab %>% 
  count(mrn, DayRT)

fgroup_relab %>% 
  write_csv('../data/finalized/all_diet_sample_with_food_group_info.csv')
```
```{r summarystats}
# how many patients how many days of data
npt <- fgroup_relab %>% 
  count(mrn) %>% 
  nrow()

pt_days <- fgroup_relab %>% 
  count(mrn, DayRT) %>% 
  ungroup() %>% 
  count(mrn)

pt_days %>% 
  gghistogram(x = 'n', bins = 30, color = 'white', fill = 'navy',
              xlab = 'Number of days of records per patient',
              title = 'How many days of records each patient has')

max(pt_days$DayRT)
min(pt_days$DayRT)
```

```{r}
# create a df to add a vertical line as the engraftment date
engraf_df <- engra %>% 
  select(mrn, dayEngra)

engraf_df %>% 
  write_csv('../data/finalized/engraf_day_df.csv')
```


```{r food_group_plot}
res <- ggplot(data = fgroup_relab, aes(x = DayRT, y = grp_relab, fill =  Main.food.description)) +
  geom_bar(position="stack", stat="identity") +
  geom_vline(data = engraf_df, aes(xintercept = dayEngra), color = 'red', size=2, linetype = 'dashed') +
  geom_vline(data = TPN, aes(xintercept = dayRT), color = 'forestgreen', size=1, linetype = 'dotted') +
  scale_fill_manual(values = scale_fill_manual_value) +
  labs(y = 'Relative Abundance') +
  #scale_x_continuous(expand = c(0, 0)) +
  #scale_y_continuous(limits = c(0, 1),expand = c(0, 0)) +
  theme_classic() +
  theme(axis.title.x=element_blank(),
        #axis.text.x=element_blank(),
        axis.ticks.x = element_blank(),
        legend.position="bottom") +
  facet_wrap(~mrn) +
  ggsave('../figs/food_group.jpg', width = 25, height = 15)
```


```{r food_group_plot}
fgroup_abs <-  ggplot(data = fgroup_relab,aes(x = DayRT, y = grp_tol, fill =  Main.food.description)) +
  geom_bar(position="stack", stat="identity") +
  geom_vline(data = engraf_df, aes(xintercept = dayEngra), color = 'red', size=2, linetype = 'dashed') +
  geom_vline(data = TPN, aes(xintercept = dayRT), color = 'forestgreen', size=1, linetype = 'dotted') +
  scale_fill_manual(values = scale_fill_manual_value) +
  labs(y = 'Total dehydrated weight in each group') +
  #scale_x_continuous(expand = c(0, 0)) +
  #scale_y_continuous(limits = c(0, 1),expand = c(0, 0)) +
  theme_classic() +
  theme(axis.title.x=element_blank(),
        #axis.text.x=element_blank(),
        axis.ticks.x = element_blank(),
        legend.position="bottom") +
  facet_wrap(~mrn) +
  ggsave('../figs/food_group_abs.jpg', width = 25, height = 15)
```

```{r}
fgroup_relab %>% 
  write_csv('../data/finalized/food_group_daily_all.csv')
```



```{r}
# pick two patients as examples and put them side by side in a large plot
patients_abs <- fgroup_relab %>% 
  filter(mrn %in% c(35522948, 35571968)) %>% 
  select(-daily_tol) %>% 
  ggplot(aes(x = DayRT, y = grp_tol, fill =  Main.food.description))+
  geom_bar(position="stack", stat="identity") +
  geom_vline(data = engraf_df %>% filter(mrn %in% c(35522948, 35571968)), aes(xintercept = dayEngra), color = 'red', size=2, linetype = 'dashed') +
  geom_vline(data = TPN %>% filter(mrn %in% c(35522948, 35571968)), aes(xintercept = dayRT), color = 'forestgreen', size=1, linetype = 'dotted') +
  labs(y = 'Total dehydrated weight in each group') +
  scale_fill_manual(values = scale_fill_manual_value) +
  theme_classic() +
  theme(axis.title.x=element_blank(),
        #axis.text.x=element_blank(),
        axis.ticks.x = element_blank(),
        legend.position="bottom") +
  facet_wrap(mrn ~ .) +
  ggsave('../figs/food_group_patients_abs.jpg', width = 12, height = 7)

patients_relab <- fgroup_relab %>% 
  filter(mrn %in% c(35522948, 35571968)) %>% 
  select(-daily_tol) %>% 
  ggplot(aes(x = DayRT, y = grp_relab, fill =  Main.food.description))+
  geom_bar(position="stack", stat="identity") +
  geom_vline(data = engraf_df %>% filter(mrn %in% c(35522948, 35571968)), aes(xintercept = dayEngra), color = 'red', size=2, linetype = 'dashed') +
  geom_vline(data = TPN %>% filter(mrn %in% c(35522948, 35571968)), aes(xintercept = dayRT), color = 'forestgreen', size=1, linetype = 'dotted') +
  labs(y = 'Relative abundance in each group') +
  scale_fill_manual(values = scale_fill_manual_value) +
  theme_classic() +
  theme(axis.title.x=element_blank(),
        #axis.text.x=element_blank(),
        axis.ticks.x = element_blank(),
        legend.position="bottom") +
  facet_wrap(mrn ~ .) +
  ggsave('../figs/food_group_patients_realb.jpg', width = 12, height = 7)
```


## The nutrient level

```{r}
colnames(Final)
```

```{r}
nutrients <- Final %>% 
  select(mrn, DayRT, Food_NSC, Calories_kcal, Proteing_g:Sodium_mg, description, dehydrated_weight) %>% 
  # CURRENTLY REMOVE THE ROWS WITH CALORIES AND TOTAL WEIGHT EQUAL TO 0
  filter(Calories_kcal != 0 & dehydrated_weight!= 0) 
```


```{r}
# get the total gram of nutrient per patient per day
nut_daily_tol <- nutrients %>% 
  select(mrn, DayRT, Proteing_g:Sodium_mg) %>% 
  gather('nutrient','weight', Proteing_g:Sodium_mg) %>% 
  group_by(mrn, DayRT,nutrient) %>% 
  summarise(nutrient_tol_day = sum(weight)) %>% 
  # conver the mg in sodium to g
  mutate(nutrient = str_replace(nutrient, '_.+$','')) %>% 
  mutate(nutrient_tol_day = if_else(nutrient == 'Sodium', nutrient_tol_day/1000, nutrient_tol_day)) %>% 
  ungroup()

# the denominator for the above data's normalization is the total weight of the 6 nutrients
nut_daily_tol_all <- nut_daily_tol %>% 
  group_by(mrn, DayRT) %>% 
  summarise(all_tol_day = sum(nutrient_tol_day))

# join them together
nutr_joined <- nut_daily_tol %>% 
  left_join(nut_daily_tol_all) %>% 
  mutate(nut_relab = nutrient_tol_day/all_tol_day) %>% 
  mutate(nutrient = factor(nutrient))

nutr_joined %>% 
  count(mrn, DayRT)

nutr_joined %>% 
  write_csv('../data/finalized/all_diet_sample_with_nutrients_info.csv')
```

```{r nutrient_plot}
# one patient for example
nutr_joined %>% 
  filter(mrn == 35586702) %>% 
  ggplot(aes(x = DayRT, y = nut_relab, fill =  nutrient)) +
  geom_bar(position="stack", stat="identity") +
  geom_vline(data = engraf_df %>% filter(mrn == 35586702), aes(xintercept = dayEngra), color = 'red', size=2, linetype = 'dashed') +
  geom_vline(data = TPN %>% filter(mrn == 35586702), aes(xintercept = dayRT), color = 'forestgreen', size=1, linetype = 'dotted') +
  scale_fill_lancet() +
  labs(y = 'Relative abundance') +
  theme_classic() +
  theme(axis.title.x=element_blank(),
        #axis.text.x=element_blank(),
        axis.ticks.x = element_blank(),
        legend.position="bottom") +
  ggsave('../figs/food_nutrient_example.jpg', dpi = 300, width = 8, height = 5)


nutr_joined %>% 
  write_csv('../data/finalized/nutrient_daily_all.csv')
```
 
```{r nutrient_plot}
nutrient_plot_relab <-  ggplot(data = nutr_joined,aes(x = DayRT, y = nut_relab, fill =  nutrient)) +
  geom_bar(position="stack", stat="identity") +
  geom_vline(data = engraf_df, aes(xintercept = dayEngra), color = 'red', size=2, linetype = 'dashed') +
  geom_vline(data = TPN, aes(xintercept = dayRT), color = 'forestgreen', size=1, linetype = 'dotted') +
  scale_fill_lancet() +
  labs(y = 'Relative abundance') +
  theme_classic() +
  theme(axis.title.x=element_blank(),
        #axis.text.x=element_blank(),
        axis.ticks.x = element_blank(),
        legend.position="bottom") +
  facet_wrap(~mrn) +
  ggsave('../figs/food_nutrient_all.jpg', width = 25, height = 15, dpi = 300)
```
