==> scripts/food_tree/scripts/05_food_diversity.Rmd <==
---
title: "Food diversity"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(ggpubr)
```




```{r}
# explaining with the nodes
NODES <- read_tsv('../data/source/NodeLabelsMCT.txt') %>% 
  mutate(Level.code= as.character(Level.code)) %>% 
  mutate(grp = str_sub(Level.code, 1,1)) %>% 
  count(grp)
bagel <- final %>% 
  mutate(Food_code= as.character(Food_code)) %>% 
  mutate(grp = str_sub(Food_code, 1,1)) %>% 
  count(grp)
  
eggs <- read_tsv('../data/source/NodeLabelsMCT.txt') %>% 
  mutate(Level.code= as.character(Level.code)) %>% 
  filter(str_detect(Level.code, '^3'))

eggs %>% 
  kableExtra::kable() %>% 
  kableExtra::kable_styling(full_width = F) %>% 
  kableExtra::save_kable('../data/source/sampleeggs.html', self_contained = T)
  
test <- final %>% 
  filter(Food_code < 40000000 & Food_code > 30000000) %>% 
  select(Food_NSC,Food_code )
```


## Alpha diversity

```{r}
final <- read_csv('../data/finalized/all_patients_record_w_food_code.csv')


# the 0 caleries items
cal0 <-  final %>% 
  filter(Calories_kcal == 0)

cal0 %>% 
  write_csv('../data/source/all_0_kcal_foods.csv')


# need to have the main food description (without special characters) as the feature name
# use the table previously generated
food_db <- read_tsv('../data/source/food_database_file.tsv')

food_cts <- final %>% 
  select(mrn, DayRT, Food_NSC, dehydrated_weight) %>% 
  left_join(food_db %>% 
              select(Food_NSC = item1_value_raw_raw,
                     Main.food.description), by  = 'Food_NSC') %>% 
  group_by(mrn,DayRT, Main.food.description) %>% 
  summarise(daily_de_wt = sum(dehydrated_weight)) %>% 
  ungroup() %>% 
  mutate(mrn = as.character(mrn),
         DayRT = as.character(DayRT)) %>% 
  mutate(DayRT = str_replace(DayRT, '-','_')) %>% 
  mutate(mrnDay = str_glue('{mrn}d{DayRT}')) %>% 
  select(-mrn, -DayRT) %>% 
  spread(key = 'mrnDay', value = 'daily_de_wt', fill = 0) %>% 
  rename(main_food_desc = Main.food.description) %>% 
  mutate(main_food_desc = str_replace_all(main_food_desc, '_',' '))

food_cts %>% 
  write_tsv('../data/finalized/all_patients_food_counts_table_w_maindescrip_featurename.tsv')

```

Convert the counts table to biom format and then to qza format, the feature names need to be the main food description

```{bash}
biom convert -i all_patients_food_counts_table_w_maindescrip_featurename.tsv -o all_patients_food_counts_table_w_maindescrip_featurename.biom --to-hdf5 --table-type="Table"
```

```{bash}
qiime tools import \
  --input-path all_patients_food_counts_table_w_maindescrip_featurename.biom \
  --output-path all_patients_food_counts_table_w_maindescrip_featurename.qza \
  --type 'FeatureTable[Frequency]'
```


Convert the tree file to qza format too

```{bash}
qiime tools import \
  --input-path output_food_tree_datatree.newick \
  --output-path output_food_tree_datatree.qza \
  --type 'Phylogeny[Rooted]'
```

Use qiime and the above two files to calculate the alpha diversity with phylogenetics


```{bash}
qiime diversity alpha-phylogenetic \
  --i-table all_patients_food_counts_table_w_maindescrip_featurename.qza \
  --i-phylogeny ../source/output_food_tree_datatree.qza \
  --p-metric faith_pd \
  --o-alpha-diversity faith_pd_vector.qza
```

```{bash}
# export to tsv format
qiime tools export --input-path faith_pd_vector.qza --output-path alpha
```


```{r}
# the results from above  
faith <- read_tsv('../data/finalized/alpha/alpha-diversity.tsv')  %>% 
  rename(mrnDay = X1) %>% 
  separate(mrnDay, into = c('mrn','DayRT'), sep = 'd') %>% 
  mutate(DayRT = str_replace(DayRT,'_','-'),
         DayRT = as.numeric(DayRT))

faith %>% 
  write_csv('../data/finalized/alpha_faith_daily_all.csv')
```

```{r}
# some phenotype info regarding the transplant
allo <- get_data_from_query_OTU(0, 'patient_allo_ag') %>% 
  filter(mrn %in% faith$mrn)
colnames(allo)

pts_info <- allo %>% 
  select(mrn, intensity, source, indication, gvhd:grade, vital_status) %>% 
  filter(str_detect(indication, 'Initial')) 

pts_info %>% 
  write_csv('../data/finalized/pheno_67.csv')
```

```{r plot_daily_alpha_div}
# plot daily alpha diversity change for every patient
faith %>% 
  ggplot(aes(x = DayRT, y = faith_pd))+
  geom_line()+
  geom_point(size = 0.8)+
  theme_classic() +
  theme(axis.title.x=element_blank(),
        #axis.text.x=element_blank(),
        axis.ticks.x = element_blank(),
        legend.position="bottom") +
  facet_wrap(~mrn)+
  ggsave('../figs/alpha_div_daily_all.jpg', width = 25, height = 15, dpi = 300)

```


```{r}
alpha_div <- faith %>% 
  full_join(pts_info %>% 
              select(mrn, vital_status))

alpha_div %>% 
  ggboxplot(x  = 'vital_status', y = 'faith_pd', add = 'jitter', 
            color = 'vital_status', palette = c('red','black'),
            xlab = '', title = 'alpha phylogenetics diversity patient wise') +
  stat_compare_means(comparisons = list(c('Alive', 'Dead')),
  label = "p.signif",
  method = 'wilcox.test',
  correct=FALSE)
```

```{r}
# the alpha diversity from the db
source('~/db.R')
dbalpha <- get_data_from_query_OTU(0, 'asv_alpha_diversity_ag')
castori <- get_data_from_query_OTU(0, 'samples_castori_ag')
# but the faith diversity is per patient basis not per sample

# get the stool sample sampleids
samps_df <- castori %>% 
  filter(mrn %in% faith$mrn)

samps <- samps_df %>% 
  pull(sampleid)

samps_alpha <- dbalpha %>% 
  filter(sampleid %in% samps)

samps_alpha_med <- samps_alpha %>% 
  left_join(samps_df %>% 
              select(sampleid, mrn)) %>% 
  group_by(mrn) %>% 
  summarise(med_simp = median(simpson_reciprocal))

stool_alpha <- samps_df %>% 
  select(datecollection, sampleid, mrn) %>% 
  inner_join(dbalpha %>% 
               select(sampleid, simpson_reciprocal)) %>% 
  left_join(allo %>% 
              select(mrn, hct)) %>% 
  mutate(dayRT = datecollection - hct) %>% 
  filter(dayRT >= -10 & dayRT <= 41)


```

```{r}
# correlation between faith diversity and simpsons diversity
cor_df <- samps_alpha_med %>% 
  full_join(alpha_div) 

pearson <- cor.test(cor_df$med_simp, cor_df$faith_pd, method = 'pearson')$estimate

spearman <- cor.test(cor_df$med_simp, cor_df$faith_pd, method = 'spearman', exact = F)$estimate
```


```{r}
# one patient with the lowerst food diversity and micriobiome diversity and one patient with high high
daily_food <- read_csv('../data/finalized/alpha_faith_daily_all.csv')

daily_food %>% 
  ggline('DayRT', 'faith_pd', title = 'Food') +
  facet_wrap(mrn ~.) +
  ggsave('../figs/two_pt.jpg', width = 30, height = 20)



stool_alpha %>% 
  #ggline('DayRT', 'faith_pd') +
  ggline(x = 'dayRT', y = 'simpson_reciprocal', color = 'red', title = 'Micriobiome') +
  facet_wrap(mrn ~.) +
  ggsave('../figs/two_pt_stool.jpg', width = 30, height = 20)


  
```
