==> scripts/food_tree/scripts/08_beta_visualization.Rmd <==
---
title: "FOOD diversity visualization: alpha and beta"
author: "Anqi Dai"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_float: true
    toc_depth: 3
    code_folding: hide
---

```{r setup, include=FALSE, out.width='150%'}
knitr::opts_chunk$set(echo = TRUE, message = F, warning = F)
```

```{r}
library(tidyverse)
library(ape)
library(vegan)
library(ggpubr)
```

```{r}
# load the pheno data of the patients 
pheno <- read_csv('../data/pheno/final_pheno.csv')
```

# beta diversity (unweighted unifrac)

```{r}
#Check the beta distance matrix
beta_un_uni <- read.delim('../data/finalized/unweighted_unifrac/distance-matrix.tsv', sep = '\t', row.names = 1)
food_dist <- as.dist(beta_un_uni)

meta <- read_tsv('../data/finalized/meta_data_67.tsv') %>% 
  rename(did = names(.)[1])

#pcoa
pcoa_f <- as.data.frame(pcoa(food_dist)$vectors) %>% 
  select(Axis.1, Axis.2) %>% 
  rownames_to_column('did') %>% 
  full_join(meta, by = 'did')


# the percent variance
eigen <- pcoa(food_dist)$values$Eigenvalues

percent_var <- signif(eigen/sum(eigen), 3)*100
```

```{r}
# look at how to set customed breaks for the viridis color gradient
pcoa_f %>% 
  gghistogram(x = 'faith_pd', bins = 200, fill = 'forestgreen')
```


```{r}
# look at the quantiles
brks <- unique(c(seq(0, 0.4, length.out = 3),seq(0.5, 0.8, length.out = 3),1.0))
brks <- seq(0, 1, 0.1)
quants_faith <- brks %>% 
  set_names(brks) %>% 
  map_dfr(~ quantile(pcoa_f$faith_pd, .)) %>% 
  gather('brks', 'value') %>% 
  mutate(value =  round(value, 0))
```


```{r}
# color by faith pd
library(viridis)
library(RColorBrewer)

pcoa_f %>% 
  mutate(brks = cut(faith_pd, 
                    breaks = format(quants_faith %>% pull(value), scientific=F),
                    include.lowest = T)) %>% 
  ggplot(aes(x = Axis.1, y = Axis.2, col = brks)) +
  geom_point(alpha=0.7, size = 1.8) +
  xlab(paste0("PC 1 [",percent_var[1],"%]")) +
  ylab(paste0("PC 2 [",percent_var[2],"%]")) +
  theme_classic() +
  labs(title = '') + 
  theme(legend.position = 'none') +
  scale_color_gradient2() +
  scale_colour_manual(values = rev(brewer.pal(10,"RdYlBu"))) +
  coord_fixed(ratio = 1) +
  #scale_color_gradient(low="royalblue", high="orangered") +
  ggsave('../figs/beta_un_unifrac_all_pts_by_faith.pdf', width = 4, height = 4)
```

```{r}
# create panels and color them by the food groups 
food_groups_sub <- pcoa_f %>% 
  select(did, Axis.1, Axis.2, foodDayRT, egg_total,egg_perc, milk_total, milk_perc, fruit_total, fruit_perc,sugar_total, sugar_perc, grain_perc, grain_total, meat_total, meat_perc, veggie_total, veggie_perc,legume_total, legume_perc, fats_perc, fats_total)

food_groups_sub_total <- food_groups_sub %>% 
  select(did, Axis.1, Axis.2, ends_with('total')) %>% 
  gather('food_group_total', 'dedywt', names(.)[4]:names(.)[ncol(.)]) 

colnames(pcoa_f)

# define the breaks
brks <- unique(c(seq(0, 0.7, length.out = 2),seq(0.8, 1, length.out = 9),1.0))
#brks <- seq(0, 1, 0.1)
quants <- brks %>% 
  set_names(brks) %>% 
  map_dfr(~ quantile(food_groups_sub_total$dedywt, .)) %>% 
  gather('brks', 'value') %>% 
  mutate(value =  round(value, 0))


food_groups_sub_total %>% 
  mutate(brks = cut(dedywt, 
                    breaks = format(quants %>% pull(value), scientific=F),
                    include.lowest = T)) %>% 
  ggplot(aes(x = Axis.1, y = Axis.2, col = brks)) +
  geom_point(alpha=0.6, size = 2.5) +
  xlab(paste0("PC 1 [",percent_var[1],"%]")) +
  ylab(paste0("PC 2 [",percent_var[2],"%]")) +
  theme_classic() +
  theme(legend.position = 'top') +
  labs(title = 'PCoA of unweighted unifrac distance colored by') + 
  scale_colour_manual(values = rev(brewer.pal(10,"RdYlBu"))) +
  facet_wrap(~food_group_total) +
  ggsave('../figs/pcoa_food_group_total.jpg', height = 10, width = 15)
```

 



```{r}
# join the coords with the pheno
beta_un_uni_df <- pcoa_f %>% 
  left_join(pheno, by  = 'mrn')

final %>% 
  distinct(Food_NSC)
```

```{r}
# look at the distribution of the DayRT cuz have fewer large values
beta_un_uni_df %>% 
  gghistogram('DayRT', bins = 120, fill = 'forestgreen', color = 'white',
              title = 'Distribution of day relative to transplant of the food records',
              xlab = 'Day relative to transplant')
```

##### More records were collected on earlier days. As the patients engrafted, they are discharged and then we don't have food records.



## PCoA plot for all patients in all time points

### Colored by day relative to transplant


```{r}
# look at the quantiles of the dayRT
quants <- seq(0, 1, 0.1)
quants <- unique(c(seq(0, 0.3, length.out = 5),seq(0.3, 0.7, length.out = 3),1.0))
#quants <- seq(0, 1, 0.05)
quantile_for_day <- quants %>% 
  set_names(quants) %>% 
  map_dfr(~quantile(beta_un_uni_df$DayRT, .)) %>% 
  gather('quantiles', 'day')



# experiment with different coloring scheme 
# define customized break that will be linked to specific colors (the breaks can be uneven to display the spread and structure of the data)
beta_un_uni_df %>% 
  mutate(brks = cut(DayRT, 
                    breaks = c(quantile_for_day %>% pull(day)),
                    #labels = c('-10:-5','-4:-2','-1:0','1:3','4:6','7:9','10:12','13:16','17:21','22:41'), 
                    include.lowest = T)) %>% 
  ggplot( aes(x = Axis.1, y = Axis.2, col = brks)) +
  geom_point(alpha=0.7, size = 2.7) +
  xlab(paste0("PC 1 [",percent_var[1],"%]")) +
  ylab(paste0("PC 2 [",percent_var[2],"%]")) +
  theme_classic() +
  theme(legend.position = 'top') +
  labs(title = 'PCoA unifrac distance colored by transplant day') +
  scale_color_brewer(palette = "RdYlGn") +
  ggsave('../figs/beta_un_unifrac_all_pts_DayRT_uneven_time_period.jpg', width = 5, height = 4)

```

##### We can see that the earlier days food data (the red-ish points) is more in the left island than in the right island. And according to the alpha diversity plot, the earlier days food data is more diverse than the later time points.

The percent variance explained by the first two PCs are relatively low, but that's in concordance with other publication, such as [Dan Knight's lab diet paper] (https://www.cell.com/cell-host-microbe/fulltext/S1931-3128(19)30250-1?_returnURL=https%3A%2F%2Flinkinghub.elsevier.com%2Fretrieve%2Fpii%2FS1931312819302501%3Fshowall%3Dtrue)



```{r}
# include the earlier time points records only
beta_un_uni_df %>% 
  filter(DayRT %in% -10:15) %>% 
  mutate(mrn = factor(as.character(mrn))) %>% 
ggplot( aes(x = Axis.1, y = Axis.2,  col = DayRT)) +
  geom_point(alpha=0.9, size = 2) +
  #stat_ellipse(color = "dark grey", type = "norm", linetype = 2, level = 0.95, alpha = 0.5) +
  theme_classic() +
  theme(legend.position = 'top') +
  scale_color_gradient(low = "yellow", high = "darkblue")  +
  labs(title = 'PCoA of unifrac distance between day -10 to day 15') + 
  facet_wrap(intensity~source) +
  ggsave('../figs/beta_un_unifrac_all_pts_DayRT_early.jpg', width = 12, height = 8)
```

##### If we zoom on to days between -10 and 15, the separation in transplant day is more clear in ablative + TCD patients.



### Color/split by intensity and source 

```{r}
# plot mark the source and the intensity

beta_un_uni_df %>% 
  mutate(mrn = factor(as.character(mrn))) %>%
  ggscatter('Axis.1', 'Axis.2', color = 'intensity', alpha = 0.4, size = 2.7,
            palette = 'lancet', title = 'PCoA unifrac distance colored by intensity') +
  facet_wrap(~source) +
  ggsave('../figs/beta_un_unifrac_all_pts_source_intensity.jpg', width = 13, height = 8)
```

- In the unmodified panel, looks like the nonablative (red) patients food records are more in the left island (the more diverse ones )
- In the TCD panel, more data clustered in the right island (the less diverse ones in the later time points). It alignes with our observation in the alpha diversity figure, that the TCD patients have a bigger drop in the diversity.



### Statistical testing on the Axis 2 before and after day -2 of all the data

```{r}
beta_un_uni_df %>% 
  mutate(grp = if_else(DayRT <= -2, 'before', 'after')) %>% 
  mutate(grp = factor(grp, levels = c('before','after'))) %>% 
  ggboxplot(x = 'grp', y = 'Axis.2', add = 'jitter', color = 'grp', palette = 'nejm',
            title = 'Axis 2 unifrac distance PCOA between before and after day -2') +
  stat_compare_means(comparisons = list(c('before', 'after')),
                      label = "p.signif",
                      method = 'wilcox.test',
                      correct=FALSE)
```

##### Using day -2 since that's the earliest abx time points. Although it's not visually obvious from the PCoA plot, the statistics show that the food composition is significantly different between the two time period in the axis 2.
