==> scripts/food_tree/scripts/10_food_alpha_GEE.Rmd <==
---
title: "diet alpha GEE"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(ggpubr)
library(geepack)
```


- This model tells the popolation difference and doesn't assume distribution of the data, and also account for multiple observations for one subject
- time^2 quadratic time cuz the change is not linear, it has U turn and such, so it's quadratic
- When you include quadratic time you also include the normal time
- the P value from summary gave you significance level
- it doesn't come together with the lowess curve, so just describe it in the text
- use the indenpent working correlations

```{r}
# load the faith alpha data
faith <- read_tsv('../data/finalized/meta_data_67.tsv')

formular_source <- formula(faith_pd ~ foodDayRT + foodDayRT^2 + source)

formular_intensity <- formula(faith_pd ~ foodDayRT + foodDayRT^2 + intensity)

# faith %>% 
#   ggboxplot(y = 'faith_pd', add = 'jitter')
```
```{r}
# all of them , without splitting
faith %>% 
  ggscatter('foodDayRT', 'faith_pd', 
             alpha = 0.8,size = 1,
            #ellipse = TRUE, mean.point = TRUE, star.plot = TRUE,
            ylab = 'alpha diversity (Faith)', xlab = 'Day relative to transplant',
            title = 'Food diversity in whole population',  
            add = "loess", 
            add.params = list(color = "blue", fill = "black"),
            conf.int = TRUE)  +
  ggsave('../figs/alpha_div_all.jpg', width = 5, height = 5)
```


```{r}
# visualization
faith %>% 
  ggscatter('foodDayRT', 'faith_pd', 
            color = 'intensity', palette = 'lancet', alpha = 0.8,size = 1,
            #ellipse = TRUE, mean.point = TRUE, star.plot = TRUE,
            ylab = 'alpha diversity (Faith)', xlab = 'Day relative to transplant',
            title = 'Food diversity by conditioning intensity',
            add = "loess", conf.int = TRUE)  +
  theme(legend.position = 'none') +
  ggsave('../figs/alpha_div_by_intensity.jpg', width = 5, height = 5)


faith %>% 
  ggscatter('foodDayRT', 'faith_pd', 
            color = 'source', palette = 'nejm', alpha = 0.8, size = 1,
            #ellipse = TRUE, mean.point = TRUE, star.plot = TRUE,
            ylab = 'alpha diversity (Faith)', xlab = 'Day relative to transplant',
            title = 'Food alpha diversity by source',
            add = "loess", conf.int = TRUE)
 
```




# by source

```{r}

# the indenpendent Working Covariance
geeInd_source <- geeglm(formular_source, id=mrn, data=faith, family=gaussian, corstr="ind")
summary(geeInd_source)
#anova(geeInd_source)

```

# by intensity

```{r}

# the indenpendent Working Covariance
geeInd_intensity <- geeglm(formular_intensity, id=mrn, data=faith, family=gaussian, corstr="ind")
summary(geeInd_intensity)
#anova(geeInd_intensity)
```
