==> scripts/food_tree/scripts/18_paired_patients_subset_procrustes_association.Rmd <==
---
title: "patients subset previous N day correlation"
author: "Anqi Dai"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_float: true
    toc_depth: 3
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = F, warning = F)
```

```{r}
library(tidyverse)
library(ggpubr)
library(vegan)
```

```{r}
asv_counts_ag <- read_csv('../data/asv_counts_ag_Feb17.csv')

source('~/db_connect_simple.R')
connect_database(config_file = '~/dbConfig.txt')
ANNOT <- get_table_from_database('asv_annotation_blast_ag')
```

## Use a subset of patient cohort: has to have at least 4 pairs that one stool sample is paired with previous 5 days of diet

```{r}
paired_stats <- read_csv('../data/finalized/paired/paired_group_summary_stats.csv')

# look at the p5day and select the patients that have at least 4 such pairs
subset_pts <- paired_stats %>% 
  filter(grp == 'p5day') %>% 
  filter(n >= 4) %>% 
  pull(mrn)
```

```{r}
# the all food records
food <- read_csv('../data/cleaned_diet_data/FINAL_97_with_code_all_foods_drt_with_enteralTPN.csv') %>% 
  select(mrn, Food_code, dehydrated_weight, fdrt ) %>% 
  mutate(fdrt = as.character(fdrt))
```

## Make paired count matrix for food and microbiome data. Diet data summed for the days

```{r}
# a function to loop over p2day ... p5day

pndays <- list.files('../data/finalized/paired/', pattern = 'p\\dday$')

  
loop_prior_N_day <- function(pnday){
  # the paired information
  fns_all <- list.files(str_glue('../data/finalized/paired/{pnday}/'), full.names = T)
  
all <- fns_all %>% 
    set_names(fns_all) %>% 
    map(~ read_csv(file = ., col_names = T, col_types = cols(.default = col_character()))) %>% 
    bind_rows() %>% 
    mutate(mrn = as.numeric(mrn)) %>% 
    filter(mrn %in% subset_pts) %>% 
    mutate(ID = str_glue('{mrn}_{sampleid}'))
```


```{r}
# for every mrn every qualified stool sample I want a sum of the previous N days dehydrated weitght for that food code
  pair_food <- apply(all, 1, function(Row){
    if(pnday == 'p5day'){
      df = food %>% 
        filter(mrn == Row[['mrn']]) %>% 
        filter(Fdrt %in% c(Row[['p1dF']], Row[['p2dF']],Row[['p3dF']],Row[['p4dF']],Row[['p5dF']])) %>% 
        group_by(Food_code) %>% 
        summarise(sum_dewt = sum(dehydrated_weight))
    }else if (pnday == 'p4day'){
      df = food %>% 
        filter(mrn == Row[['mrn']]) %>% 
        filter(Fdrt %in% c(Row[['p1dF']], Row[['p2dF']],Row[['p3dF']],Row[['p4dF']])) %>% 
        group_by(Food_code) %>% 
        summarise(sum_dewt = sum(dehydrated_weight))
    }else if (pnday == 'p3day'){
      df = food %>% 
        filter(mrn == Row[['mrn']]) %>% 
        filter(Fdrt %in% c(Row[['p1dF']], Row[['p2dF']],Row[['p3dF']])) %>% 
        group_by(Food_code) %>% 
        summarise(sum_dewt = sum(dehydrated_weight))
    }else if (pnday == 'p2day'){
      df = food %>% 
        filter(mrn == Row[['mrn']]) %>% 
        filter(Fdrt %in% c(Row[['p1dF']], Row[['p2dF']])) %>% 
        group_by(Food_code) %>% 
        summarise(sum_dewt = sum(dehydrated_weight))
    }
  return(df)
})
  
  names(pair_food) <- all$ID
  
  pair_food_df <- pair_food %>% 
    bind_rows(.id = 'ID') %>% 
    separate(ID, into = c('mrn','sampleid'), sep = '_') 
  
  pair_pt_food <- pair_food_df %>% 
    split(.$mrn) %>% 
    map(~ select(.data = ., -mrn) %>% 
          spread(key = 'sampleid', value = 'sum_dewt', fill = 0))
  
  # get the microbiome counts data of those stool samples
  cts <- asv_counts_ag %>% 
    filter(sampleid %in% all$sampleid) %>% 
    select(asv_key, sampleid, count) %>% 
    spread(key = 'sampleid', value = 'count', fill = 0) %>% 
    arrange(asv_key)  
  
  annot <- ANNOT %>% 
    filter(asv_key %in% cts$asv_key) %>% 
    mutate(ordr =  if_else(ordr == '', str_glue('of_class_{class}'), ordr),
           family =  if_else(family == '', str_glue('of_order_{ordr}'), family),
           genus =  if_else(genus == '', str_glue('of_family_{family}'), genus),
           species =  if_else(species == '', str_glue('of_genus_{genus}'), species)) %>% 
    mutate(taxa_species = str_glue('k__{kingdom}|p__{phylum}|c__{class}|o__{ordr}|f__{family}|g__{genus}|s__{species}'))
  
  # replace the asv_key with the genus taxa level
  # summarize from asv level to genus level
  # get the relative abundance for the species
  cts_spilt_pt <- cts %>% 
    full_join(annot %>%  select(asv_key, taxa_species), by  = 'asv_key') %>% 
    select(-asv_key) %>% 
    gather(key = 'sampleid', value = 'count', names(.)[1]:names(.)[ncol(.) - 1]) %>% 
    group_by(sampleid, taxa_species) %>% 
    summarise(cnt = sum(count)) %>% 
    # get the total count from the db to calculate the relab
    left_join(asv_counts_ag %>% distinct(sampleid,count_total ), by = 'sampleid') %>% 
    mutate(relab = cnt/count_total) %>% 
    select(sampleid, taxa_species, relab) %>% 
    left_join(all %>% select(mrn, sampleid), by  = 'sampleid') %>% 
    arrange(mrn, sampleid) %>% 
    split(.$mrn) %>% 
    map(~ select(.data = ., -mrn) %>% 
          spread(key = 'sampleid', value = 'relab'))
  
  # output the files to be run thru snakemake using the qiime
  # for the stool
  for(pt in as.character(subset_pts)){
    cts_spilt_pt %>% 
      pluck(pt) %>% 
      write_tsv(str_glue('../data/finalized/paired/pNday/{pnday}_{pt}_stool_relab_species.tsv'))
  }
  
  names_loop <- data_frame(
    name = as.character(subset_pts) 
  ) %>% 
      transmute(name = str_glue('{pnday}_{name}')) %>% 
    write_csv(str_glue('../data/finalized/paired/pNday/{pnday}_names_loop.csv'), col_names = F)
  
  # for the food
  for(pt in as.character(subset_pts)){
    pair_pt_food %>% 
      pluck(pt) %>% 
      write_tsv(str_glue('../data/finalized/paired/pNday/{pnday}_{pt}_diet_foodID_dehydrated_weight_per_pt.tsv'))
}
}


pndays %>% 
  map(~ loop_prior_N_day(pnday = .))

```

## Using the PCOA produced by qiime to do the procrustes analysis 

```{r}
# after getting the results from qiime 
set.seed(123)

# sort out the fns that I need to do a procrustes on
fns_both <- data_frame(
  sfolder = list.files('../data/finalized/paired/pNday/', pattern = 'p.day.+_stool_relab_species_braycurtis_pcoa$', full.names = T),
  dfolder = list.files('../data/finalized/paired/pNday/', pattern = 'p.day.+_diet_foodID_dehydrated_weight_per_pt_unweighted_unifrac_pcoa$', full.names = T)
) %>% 
  mutate(ID = str_replace(sfolder, '_stool_relab_species_braycurtis_pcoa', ''),
         ID = str_replace(ID, '../data/finalized/paired/pNday//','')) %>% 
  separate(ID, into = c('pNday','mrn'), sep = '_') %>% 
  mutate(sfn = str_glue('{sfolder}/ordination.txt'),
         dfn = str_glue('{dfolder}/ordination.txt'))


# a function to do the procrustes as a pair
noprocrustes_Nday <- map2_dfc(fns_both$dfn, fns_both$sfn, function(dfn, sfn){
  dpcoa = read_tsv(dfn, skip = 9, col_names = F)  %>% 
    filter(! X1 %in% c('Biplot','Site constraints')) %>% 
    column_to_rownames('X1') %>% 
    as.matrix()
  
  spcoa = read_tsv(sfn, skip = 9, col_names = F)  %>% 
    filter(! X1 %in% c('Biplot','Site constraints')) %>% 
    column_to_rownames('X1') %>% 
    as.matrix()
  
  protest_res = protest(X = dpcoa, Y = spcoa,  permutations = 999)
  return( list(corr = protest_res$t0,
               pval = protest_res$signif))
}) %>% 
  gather(key = 'mrn', value = 'corre') %>% 
  mutate(mrn = fns_both$mrn,
         pNday = fns_both$pNday) %>% 
  arrange(corre)
```



## Visualize the results

```{r}
# now visualize the correlation instead of the p value
procrustes_Nday %>% 
  mutate(day = str_extract(pNday, pattern = '\\d')) %>% 
  mutate(day = factor(day)) %>% 
  ggboxplot(x = 'day', y = 'corre', 
            ylab = 'Correlation in a symmetric Procrustes rotation', xlab = 'Number of days combined',
            title = '') +
  #geom_jitter(aes(color = mrn), size = 5, alpha = 0.8, width = 0.05) +
  #geom_line(aes(group = mrn), color = "grey", alpha = 0.5) +
  theme(legend.position = 'none') +
  ggsave('../figs/pNday_sub_pts_boxplot_corre.pdf', width = 6, height = 4) 
```


```{r}
# plot and visualize
procrustes_Nday_df <- procrustes_Nday %>% 
  mutate(neglogPval = -log10(pval)) %>% 
  mutate(day = str_extract(pNday, pattern = '\\d')) 

procrustes_Nday_df %>% 
  ggboxplot(x = 'day', y = 'neglogPval', 
            ylab = '-log10(pval)', xlab = 'Number of days combined',
            title = '') +
  #geom_jitter(aes(color = mrn), size = 5, alpha = 0.8, width = 0.05) +
  #geom_line(aes(group = mrn), color = "grey", alpha = 0.5) +
  theme(legend.position = 'none') +
  ggsave('../figs/pNday_sub_pts_boxplot.pdf', width = 6, height = 4) 


```
