==> scripts/food_tree/scripts/22_correlate_food_w_microbiome_family_taxon.Rmd <==
---
title: "correlate food with microbiome family taxon"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(ggpubr) 
```

```{r}
source('~/db_connect_simple.R')
connect_database(config_file = '~/dbConfig.txt')
get_table_from_database_predefined_filter('asv_counts_ag')
get_table_from_database('asv_annotation_blast_ag')
```

```{r}
pair3 <- read_csv('../data/finalized/paired/pair_p3day_food_level2_res_fil.csv')

nodes <- read_tsv('../data/source/NodeLabelsMCT.txt', col_types = cols(.default = col_character())) %>% 
  filter(str_length(Level.code) == 2) %>% 
  # remove the ones ending with 0 like 20 and 50
  filter(!str_detect(Level.code, '0$'))

nodes_l1 <- read_tsv('../data/source/NodeLabelsMCT.txt', col_types = cols(.default = col_character())) %>% 
  filter(str_length(Level.code) == 1)  %>% 
  rename(food_l1 = Level.code,
         desc_l1 = Main.food.description)
```


```{r}
cts <- asv_counts_ag %>% 
  filter(sampleid %in% pair3$sampleid) %>% 
  select(asv_key, sampleid, count) %>% 
  spread(key = 'sampleid', value = 'count', fill = 0) %>% 
  arrange(asv_key)  

annot <- asv_annotation_blast_ag %>% 
  filter(asv_key %in% cts$asv_key) %>% 
  mutate(ordr =  if_else(ordr == '', str_glue('of_class_{class}'), ordr),
         family =  if_else(family == '', str_glue('of_order_{ordr}'), family),
         genus =  if_else(genus == '', str_glue('of_family_{family}'), genus),
         species =  if_else(species == '', str_glue('of_genus_{genus}'), species)) %>% 
  mutate(taxa_family = str_glue('k__{kingdom}|p__{phylum}|c__{class}|o__{ordr}|f__{family}'))


# replace the asv_key with the genus taxa level
# summarize from asv level to genus level
# get the relative abundance for the genera
cts_all <- cts %>% 
  full_join(annot %>%  select(asv_key, taxa_family), by  = 'asv_key') %>% 
  select(-asv_key) %>% 
  gather(key = 'sampleid', value = 'count', names(.)[1]:names(.)[ncol(.) - 1]) %>% 
  group_by(sampleid, taxa_family) %>% 
  summarise(cnt = sum(count)) %>% 
  # get the total count from the db to calculate the relab
  left_join(asv_counts_ag %>% distinct(sampleid,count_total ), by = 'sampleid') %>% 
  mutate(relab = cnt/count_total) %>% 
  select(sampleid, taxa_family, relab)  

cts_all_keep_feature <- cts_all %>% 
  spread(key = 'sampleid', value = 'relab', fill = 0) %>% 
  column_to_rownames('taxa_family')

# filtering:  only microbiome features that are present in >= minmum-prevlence (e.g. 10% of patients)  of the samples and above some minimum relative abundance, (e.g. >= XXX).

pass_thre_num <- ceiling(ncol(cts_all_keep_feature) * 0.1)

counts_fil_keep_feature <- rownames(cts_all_keep_feature[rowSums(cts_all_keep_feature > 0.001) >= pass_thre_num, ])

cts_all_fil <- cts_all %>% 
  filter(taxa_family %in% counts_fil_keep_feature)

pair3_both <- pair3 %>% 
  left_join(cts_all_fil , by  = 'sampleid')
```

```{r}
# out put the cts_all_fil these are the qualified samples genus relab for the genus that have relab > 0.001 in >= 10% of samples
cts_all_fil %>% 
  write_csv('../data/finalized/paired/pair_p3day_qualified_stool_sampled_family_relab.csv')

cts_all_fil %>% 
  ungroup() %>% 
  distinct(taxa_family)
```


```{r}
# every species relab and correlate with the level 2 food group ave weight

pair3_both_spearm <- pair3_both %>% 
  split(.$taxa_family) %>% 
  map(function(df){
     df %>% 
        split(.$Level.code) %>% 
        map(function(df_l2){
          test = suppressWarnings(cor.test(df_l2$ave_level2, 
                         df_l2$relab, 
                         method = 'spearman', 
                         exact = F))
          return(list(rho = round(test$estimate, 2),
                      pval = test$p.value))}) %>% 
        bind_rows(.id = 'Level.code')
  }) %>% 
  bind_rows(.id = 'taxa_family') %>% 
  filter(!is.na(rho)) %>% 
  mutate(FDR = p.adjust(pval, method = 'BH')) %>% 
  mutate(neglogFDR = -log(FDR)) %>% 
  arrange(desc(neglogFDR)) 

```

```{r}
# get the sig ones (FDR < 0.1)
pair3_both_spearm_sig <- pair3_both_spearm %>% 
  filter(FDR < 0.1) %>% 
  mutate(family = str_extract(taxa_family, 'f__.+$'))  %>% 
  # join with the node lables 
  left_join(nodes , by  = 'Level.code') %>% 
  rename(description = Main.food.description) %>% 
  #mutate(spp = str_extract(taxa_spp, 's__.+$')) %>% 
  # to add a tiny amount to avoid the FDR == 0 situation  
  mutate(FDR = if_else(FDR == 0, FDR + 10^-6, FDR)) %>% 
  mutate(neglogFDR = -log(FDR)) %>% 
  # get the L1 food group names 
  mutate(food_l1 = str_sub(Level.code, start = 1, end = 1)) %>% 
  left_join(nodes_l1, by = 'food_l1')
```

```{r}
# trying to plot like the diet paper fig4 a
  
pair3_both_spearm_sig %>% 
  ggscatter(x = 'rho', y = 'family', size = 'neglogFDR', label = 'family', repel = T,
            xlab = 'Spearman rho', ylab = 'Species name', 
            title = 'Spearman cor between stool sample family relab and previous three days diet\nlevel 2 food group dehydrated weight FDR < 0.1 result (abundance threshold > 0.01)') +
  facet_grid(desc_l1 ~ ., scales = "free", space = "free_y") +
  theme_bw() +
  geom_text(aes(x = 1.5, label = description)) +
  geom_vline(xintercept = c(0.5, -0.5), linetype = 'dashed', color = 'red') +
  geom_vline(xintercept = 1, color = 'black') +
  theme(axis.text.y=element_text(size=8),
        legend.position = 'none') +
  ggsave('../figs/pair3_both_spearman_family_taxon_food_L2_sig.pdf', height = 20, width = 15) 
```
